[feat] Normalizer and featurizer files order modified
